move pid function into LibGny.cpp
